# Add fast path for string collections