--- a/MaxMind.Db.Benchmark/MaxMind.Db.Benchmark.csproj
+++ b/MaxMind.Db.Benchmark/MaxMind.Db.Benchmark.csproj
@@ -34,4 +34,8 @@
     <ProjectReference Include="..\MaxMind.Db\MaxMind.Db.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.14.0" />
+  </ItemGroup>
+
 </Project>