Optimization that gives a ~10% boost on a partial `CityResponse`.

Here is benchmark using 23.56.162.100 from the official GeoIP2-City.mmdb.

| Method | Mean     | Error    | StdDev   | Gen0   | Allocated |
|------- |---------:|---------:|---------:|-------:|----------:|
| Before | 10.49 us | 0.206 us | 0.192 us | 0.3662 |   6.08 KB |
| After  | 9.397 us | 0.023 us | 0.019 us | 0.3350 |   5.70 KB |