@verdie-g 2025-04-26
The result is not as great when using 214.0.0.0 from GeoIP2-City-Test.mmdb.

| Method | Mean     | Error    | StdDev   | Gen0   | Allocated |
|------- |---------:|---------:|---------:|-------:|----------:|
| Before | 10.33 us | 0.026 us | 0.023 us | 0.3967 |   6.66 KB |
| After  | 10.04 us | 0.074 us | 0.069 us | 0.3815 |   6.29 KB |